% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_issue_module.R
\name{report_issue_module}
\alias{report_issue_module}
\title{Report Issue Module Server}
\usage{
report_issue_module(input, output, session, repo, gh_pat)
}
\arguments{
\item{input}{the Shiny input}

\item{output}{the Shiny output}

\item{session}{the Shiny session}

\item{repo}{the name of the repo to create the issue in.  e.g. "tychobra/lipids_study"}

\item{gh_pat}{the GitHub PAT}
}
\description{
Report Issue Module Server
}
