% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_issue_module.R
\name{report_issue_module_dash_ui}
\alias{report_issue_module_dash_ui}
\title{Report Issue Module UI}
\usage{
report_issue_module_dash_ui(id)
}
\arguments{
\item{id}{the module ID}
}
\value{
"Report Issue" button to go in the top bar of a shinydashboard Shiny app.
}
\description{
Report Issue Module UI
}
